move fixtures to be more general
