+++ tests/unit/conftest.py
@@ -10,6 +10,10 @@ import os
 import json
 import pytest
 
+from .compat import mock
+
+from ...plugins.module_utils._authenticator import HashiVaultAuthenticator
+
 
 @pytest.fixture(autouse=True)
 def skip_python():
@@ -39,3 +43,34 @@ def fixture_loader():
         return d
 
     return _loader
+
+
+@pytest.fixture
+def vault_client():
+    return mock.MagicMock()
+
+
+@pytest.fixture
+def patch_authenticator():
+    authenticator = HashiVaultAuthenticator
+    authenticator.validate = lambda self: True
+    authenticator.authenticate = lambda self, client: 'dummy'
+
+    with mock.patch('ansible_collections.community.hashi_vault.plugins.module_utils._hashi_vault_module.HashiVaultAuthenticator', new=authenticator):
+        yield
+
+
+@pytest.fixture
+def patch_get_vault_client(vault_client):
+    with mock.patch(
+        'ansible_collections.community.hashi_vault.plugins.module_utils._hashi_vault_common.HashiVaultHelper.get_vault_client', return_value=vault_client
+    ):
+        yield
+
+
+@pytest.fixture
+def requests_unparseable_response():
+    r = mock.MagicMock()
+    r.json.side_effect = json.JSONDecodeError
+
+    return r
